+++ packages/vue/src/index.ts
@@ -341,4 +341,4 @@ export type { VuePrintLayoutProps } from './components/PrintLayout'
 export { ImageViewer } from './components/ImageViewer'
 export type { VueImageViewerProps } from './components/ImageViewer'
 
-export const version = '0.0.1'
+export const version = '1.1.0'